Fix an issue with backslashes ending up in the image URL path.
Fix the _Layout.cshtml page to include the lolCat .css and .js files,
instead of including all the code in the head of the html document.

Update Index.cshtml - jQuery now obtains all the images in the database,
and writes their URLS as separate <img> tags in a list (<li>).

The slideshow then cycles through this list of items, showing one at a
time using a dissolve effect.